+++ src/main/java/com/example/notes/repository/UserRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.notes.model.Users;
 
 
-public interface UserRepo extends JpaRepository<Users, Integer> {
+public interface UserRepo extends JpaRepository<Users, Long> {
     Users findByUsername(String username);
 }
